tegra: redundant-boot-overrides: add service to replace nv_update_verifier
With recent versions of L4T, the NVIDIA bootloader on platforms
that support A/B redundancy will automatically decrement the priority
level of the current boot slot before starting the operating system,
expecting that the nv_update_verifier service will reset it if the
OS boot is successful.  If the system boots three times without
resetting the priority, the bootloader will fail over to the alternate
boot slot.

For Mender, we do not use nv_update_verifier due to some undesirable
side-effects, so we mask it from being run at startup.  To prevent
the failover from occurring, provide our own service to run the
necessary commands to both mark the current boot successful and
reset the slot priority.

Signed-off-by: Matt Madison <matt@madison.systems>